Switch over to esbuild
Things are speedy(TM) now, we had to write a tonne of plugins (copy,
html, license, postcss) to do this.

The resulting file is a chunk larger (460k vs 440k) and there is some
ugliness due to https://github.com/evanw/esbuild/issues/334, but it's
not the end of the world - worth it just for the near-instant builds.